코드 리팩토링 (#6)
- useFloating 훅 분리
- 검색결과 클릭시 본문 연결 기능 추가
- 컴포넌트 / 스타일 모듈 분리
