+++ src/utils/url.ts
@@ -1,0 +1,2 @@
+export const getSearchURL = (name: string) =>
+  `https://clinicaltrialskorea.com/studies?conditions=${encodeURI(name)}`;
